pyproject.toml
[tool.poetry]
name = "IchnaeaDevelopment"
version = "0.1.1"
description = "MPPT charge controller development environment"
authors = ["Brandon Braun <brandonbraun653@protonmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.12"
crc = "^7.0.0"
eseries = "^1.2.1"
grpcio = "^1.65.0"
grpcio-tools = "^1.65.0"
loguru = "^0.7.2"
matplotlib = "^3.8.3"
mypy-protobuf = "^3.6.0"
protobuf = "^5.27.2"
pytest = "*"
pytest-loguru = "^0.4.0"

mbedutils = {path = "lib/mbedutils/python", develop = true}
ichnaea = {path = "app", develop = true}
pytest-check = "^2.3.1"
oct2py = "^5.7.2"
numpy = "^2.1.3"
flake8 = "^7.1.1"
black = "^24.10.0"
pytest-xdist = "^3.6.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
# ^/foo.py  # exclude a file named foo.py in the root of the project
'''